@@ -3,25 +3,38 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "express": "^4.16.3",
-    "react": "^16.4.2",
-    "react-bootstrap": "^0.32.1",
-    "react-dom": "^16.4.2",
-    "react-scripts": "1.1.4",
-    "request": "^2.88.0",
-    "request-promise": "^4.2.2"
+    "bootstrap": "^4.3.1",
+    "googlephotos": "^0.1.2",
+    "react": "^16.8.6",
+    "react-bootstrap": "^1.0.0-beta.12",
+    "react-cookie": "^4.0.0",
+    "react-dom": "^16.8.6",
+    "react-google-login": "^5.0.5",
+    "react-router-dom": "^4.3.1",
+    "react-scripts": "3.0.0"
   },
   "homepage": "http://anandkumarpatel.github.io/smile-finder",
   "scripts": {
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build",
-    "start": "node server/server.js",
-    "demo": "react-scripts start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
   },
   "devDependencies": {
-    "gh-pages": "^1.2.0"
+    "gh-pages": "^2.0.1"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
